restructure course preview folder
