+++ src/pages/CoursePreview/_components/CoursePreviewText.tsx
