+++ src/pages/CoursePreview/_components/CoursePreviewVideo.tsx
