+++ src/pages/index.ts
@@ -1,3 +1,4 @@
+import CoursePreview from './CoursePreview'
 import { default as ManagerOverview, default as Overview } from './Overview'
 import PaymentStatus from './PaymentStatus'
 import Pricing from './Pricing'
@@ -6,7 +7,6 @@ import SignUp from './SignUp'
 import Courses from './manager/Courses'
 import CreateCourse from './manager/Courses/CreateCourse'
 import CourseDetail from './manager/Courses/ManageCourse'
-import CoursePreview from './manager/Courses/ManageCourse/CoursePreview'
 import CreateContent from './manager/Courses/ManageCourse/CreateContent'
 import Students from './manager/Students'
